cgame: avoid some casting

+++ src/cgame/cg_view.c
@@ -1015,7 +1015,7 @@ static int CG_CalcFov(void)
 			// do smooth transitions for the binocs
 			if (cg.zoomedBinoc)         // binoc zooming in
 			{
-				f = (cg.time - cg.zoomTime) / (float)ZOOM_TIME;
+				f = (cg.time - cg.zoomTime) / ZOOM_TIME;
 				if (f > 1.0)
 				{
 					fov_x = zoomFov;
@@ -1033,7 +1033,7 @@ static int CG_CalcFov(void)
 			}
 			else                        // binoc zooming out
 			{
-				f = (cg.time - cg.zoomTime) / (float)ZOOM_TIME;
+				f = (cg.time - cg.zoomTime) / ZOOM_TIME;
 				if (f > 1.0)
 				{
 					// fov_x = fov_x;
@@ -1574,7 +1574,7 @@ void CG_DrawSkyBoxPortal(qboolean fLocalView)
 			// do smooth transitions for the binocs
 			if (cg.zoomedBinoc)            // binoc zooming in
 			{
-				f       = (cg.time - cg.zoomTime) / (float)ZOOM_TIME;
+				f       = (cg.time - cg.zoomTime) / ZOOM_TIME;
 				fov_x   = (f > 1.0) ? zoomFov : fov_x + f * (zoomFov - fov_x);
 				lastfov = fov_x;
 			}
@@ -1585,7 +1585,7 @@ void CG_DrawSkyBoxPortal(qboolean fLocalView)
 			}
 			else                        // binoc zooming out
 			{
-				f     = (cg.time - cg.zoomTime) / (float)ZOOM_TIME;
+				f     = (cg.time - cg.zoomTime) / ZOOM_TIME;
 				fov_x = (f > 1.0) ? fov_x : zoomFov + f * (fov_x - zoomFov);
 			}
 		}
